LayoutTests/fast/filesystem/file-writer-abort-continue-expected.txt

Test that FileWriter can continue immediately after an abort.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

starting test
PASS 1100000 is blobSize
PASS Calling abort
PASS Saw abort
PASS Saw writeend 0.
PASS writer.length is 0
PASS followOnAction == 1 || followOnAction == 3 is true
PASS 1100000 is blobSize
PASS writer.length is 1100000
PASS Saw writeend 1.
PASS 1100000 is blobSize
PASS Calling abort
PASS Saw abort
PASS Saw writeend 0.
PASS writer.length is 1100000
PASS followOnAction == 1 || followOnAction == 3 is true
PASS writer.length is 7
PASS All tests complete.
PASS successfullyParsed is true

TEST COMPLETE
